// src/customers/booking/dto/booking-status.dto.ts
export enum BookingStatus {
  CREATED = 'CREATED',
  VEHICLE_SELECTED = 'VEHICLE_SELECTED',
  PAYMENT_PREVIEW = 'PAYMENT_PREVIEW',
  PAYMENT_INITIATED = 'PAYMENT_INITIATED',
  CONFIRMED = 'CONFIRMED',
  DRIVER_ASSIGNED = 'DRIVER_ASSIGNED',
  DRIVER_NOT_FOUND = 'DRIVER_NOT_FOUND',
  TRIP_STARTED = 'TRIP_STARTED',
  TRIP_COMPLETED = 'TRIP_COMPLETED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
}